deployment bug fixed due to next/script
